requirements.txt
# Word 文檔處理
python-docx>=0.8.11

# 檔案操作
pathlib>=1.0.1

# HTML 處理
beautifulsoup4>=4.9.3
lxml>=4.6.3

# 翻譯服務
deepl>=1.8.0
requests>=2.28.1
googletrans>=4.0.0-rc1

# JSON 處理
jsonschema>=3.2.0

# 日期處理
python-dateutil>=2.8.2

# 日誌記錄
loguru>=0.5.3

# 進度條
tqdm>=4.62.3

# 字符串處理
ftfy>=6.0.3

# 錯誤處理
sentry-sdk>=1.5.1

# 測試
pytest>=6.2.5
pytest-cov>=2.12.1